backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
kombu==5.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
schedule==1.2.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
